text_to_sql/requirements.txt
flask
anthropic
gunicorn
python-dotenv